+++ core/src/main/java/io/cucumber/core/options/PluginOption.java
@@ -1,6 +1,5 @@
 package io.cucumber.core.options;
 
-import io.cucumber.core.exception.CucumberException;
 import io.cucumber.core.logging.Logger;
 import io.cucumber.core.logging.LoggerFactory;
 import io.cucumber.core.plugin.DefaultSummaryPrinter;
@@ -25,56 +24,64 @@ import io.cucumber.plugin.SummaryPrinter;
 
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Map;
 import java.util.Set;
-import java.util.function.Supplier;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+
+import static java.util.Collections.unmodifiableMap;
+import static java.util.Collections.unmodifiableSet;
+import static java.util.stream.Collectors.joining;
 
 public class PluginOption implements Options.Plugin {
 
     private static final Logger log = LoggerFactory.getLogger(PluginOption.class);
 
     private static final Pattern PLUGIN_WITH_ARGUMENT_PATTERN = Pattern.compile("([^:]+):(.*)");
-    private static final HashMap<String, Supplier<Class<? extends Plugin>>> PLUGIN_CLASSES = new HashMap<String, Supplier<Class<? extends Plugin>>>() {
-        {
-            put("default_summary", () -> DefaultSummaryPrinter.class);
-            put("html", () -> HtmlFormatter.class);
-            put("json", () -> JsonFormatter.class);
-            put("junit", () -> JUnitFormatter.class);
-            put("null_summary", () -> NullSummaryPrinter.class);
-            put("pretty", () -> PrettyFormatter.class);
-            put("progress", () -> ProgressFormatter.class);
-            put("message", () -> MessageFormatter.class);
-            put("rerun", () -> RerunFormatter.class);
-            put("summary", () -> DefaultSummaryPrinter.class);
-            put("testng", () -> TestNGFormatter.class);
-            put("timeline", () -> TimelineFormatter.class);
-            put("unused", () -> UnusedStepsSummaryPrinter.class);
-            put("usage", () -> UsageFormatter.class);
-            put("teamcity", () -> TeamCityPlugin.class);
-        }
-    };
+    private static final Map<String, Class<? extends Plugin>> PLUGIN_CLASSES;
 
-    // Replace IDEA plugin with TeamCity
-    private static final Set<String> INCOMPATIBLE_INTELLIJ_IDEA_PLUGIN_CLASSES = new HashSet<String>() {
-        {
-            add("org.jetbrains.plugins.cucumber.java.run.CucumberJvmSMFormatter");
-            add("org.jetbrains.plugins.cucumber.java.run.CucumberJvm2SMFormatter");
-            add("org.jetbrains.plugins.cucumber.java.run.CucumberJvm3SMFormatter");
-            add("org.jetbrains.plugins.cucumber.java.run.CucumberJvm4SMFormatter");
-            add("org.jetbrains.plugins.cucumber.java.run.CucumberJvm5SMFormatter");
-        }
-    };
+    static {
+        Map<String, Class<? extends Plugin>> plugins = new HashMap<>();
+        plugins.put("default_summary", DefaultSummaryPrinter.class);
+        plugins.put("html", HtmlFormatter.class);
+        plugins.put("json", JsonFormatter.class);
+        plugins.put("junit", JUnitFormatter.class);
+        plugins.put("null_summary", NullSummaryPrinter.class);
+        plugins.put("pretty", PrettyFormatter.class);
+        plugins.put("progress", ProgressFormatter.class);
+        plugins.put("message", MessageFormatter.class);
+        plugins.put("rerun", RerunFormatter.class);
+        plugins.put("summary", DefaultSummaryPrinter.class);
+        plugins.put("testng", TestNGFormatter.class);
+        plugins.put("timeline", TimelineFormatter.class);
+        plugins.put("unused", UnusedStepsSummaryPrinter.class);
+        plugins.put("usage", UsageFormatter.class);
+        plugins.put("teamcity", TeamCityPlugin.class);
+        PLUGIN_CLASSES = unmodifiableMap(plugins);
+    }
 
-    // Refuse plugins known to implement the old API
-    private static final Set<String> INCOMPATIBLE_PLUGIN_CLASSES = new HashSet<String>() {
-        {
-            add("io.qameta.allure.cucumberjvm.AllureCucumberJvm");
-            add("io.qameta.allure.cucumber2jvm.AllureCucumber2Jvm");
-            add("io.qameta.allure.cucumber3jvm.AllureCucumber3Jvm");
-            add("io.qameta.allure.cucumber4jvm.AllureCucumber4Jvm");
-        }
-    };
+    private static final Set<String> INCOMPATIBLE_INTELLIJ_IDEA_PLUGIN_CLASSES;
+
+    static {
+        Set<String> incompatible = new HashSet<>();
+        incompatible.add("org.jetbrains.plugins.cucumber.java.run.CucumberJvmSMFormatter");
+        incompatible.add("org.jetbrains.plugins.cucumber.java.run.CucumberJvm2SMFormatter");
+        incompatible.add("org.jetbrains.plugins.cucumber.java.run.CucumberJvm3SMFormatter");
+        incompatible.add("org.jetbrains.plugins.cucumber.java.run.CucumberJvm4SMFormatter");
+        incompatible.add("org.jetbrains.plugins.cucumber.java.run.CucumberJvm5SMFormatter");
+        INCOMPATIBLE_INTELLIJ_IDEA_PLUGIN_CLASSES = unmodifiableSet(incompatible);
+    }
+
+    private static final Set<String> INCOMPATIBLE_PLUGIN_CLASSES;
+
+    static {
+        Set<String> incompatible = new HashSet<>();
+        incompatible.add("io.qameta.allure.cucumberjvm.AllureCucumberJvm");
+        incompatible.add("io.qameta.allure.cucumber2jvm.AllureCucumber2Jvm");
+        incompatible.add("io.qameta.allure.cucumber3jvm.AllureCucumber3Jvm");
+        incompatible.add("io.qameta.allure.cucumber4jvm.AllureCucumber4Jvm");
+        INCOMPATIBLE_PLUGIN_CLASSES = unmodifiableSet(incompatible);
+    }
 
     private final String pluginString;
     private final Class<? extends Plugin> pluginClass;
@@ -86,41 +93,79 @@ public class PluginOption implements Options.Plugin {
         this.argument = argument;
     }
 
-    public static PluginOption parse(String pluginArgumentPattern) {
-        Matcher pluginWithFile = PLUGIN_WITH_ARGUMENT_PATTERN.matcher(pluginArgumentPattern);
+    public static PluginOption parse(String pluginSpecification) {
+        Matcher pluginWithFile = PLUGIN_WITH_ARGUMENT_PATTERN.matcher(pluginSpecification);
         if (!pluginWithFile.matches()) {
-            return new PluginOption(pluginArgumentPattern, parsePluginName(pluginArgumentPattern), null);
+            Class<? extends Plugin> pluginClass = parsePluginName(pluginSpecification, pluginSpecification);
+            return new PluginOption(pluginSpecification, pluginClass, null);
         }
 
-        Class<? extends Plugin> pluginClass = parsePluginName(pluginWithFile.group(1));
-        return new PluginOption(pluginArgumentPattern, pluginClass, pluginWithFile.group(2));
+        Class<? extends Plugin> pluginClass = parsePluginName(pluginSpecification, pluginWithFile.group(1));
+        return new PluginOption(pluginSpecification, pluginClass, pluginWithFile.group(2));
     }
 
-    private static Class<? extends Plugin> parsePluginName(String pluginName) {
+    @SuppressWarnings("unchecked")
+    private static Class<? extends Plugin> parsePluginName(String pluginSpecification, String pluginName) {
+        // Refuse plugins known to implement the old API
         if (INCOMPATIBLE_PLUGIN_CLASSES.contains(pluginName)) {
-            throw new IllegalArgumentException("Plugin is not compatible with this version of Cucumber: " + pluginName);
+            throw createPluginIsNotCompatible(pluginSpecification);
         }
 
+        // Replace IDEA plugin with TeamCity
         if (INCOMPATIBLE_INTELLIJ_IDEA_PLUGIN_CLASSES.contains(pluginName)) {
             log.debug(() -> "Incompatible IntelliJ IDEA Plugin detected. Falling back to teamcity plugin");
             return TeamCityPlugin.class;
         }
-        return PLUGIN_CLASSES.getOrDefault(pluginName, () -> loadClass(pluginName)).get();
-    }
 
-    @SuppressWarnings("unchecked")
-    private static Class<? extends Plugin> loadClass(String className) {
+        if (PLUGIN_CLASSES.containsKey(pluginName)) {
+            return PLUGIN_CLASSES.get(pluginName);
+        }
+
         try {
-            Class<?> aClass = Thread.currentThread().getContextClassLoader().loadClass(className);
-
+            Class<?> aClass = Thread.currentThread().getContextClassLoader().loadClass(pluginName);
             if (Plugin.class.isAssignableFrom(aClass)) {
                 return (Class<? extends Plugin>) aClass;
             }
-            throw new CucumberException(
-                "Couldn't load plugin class: " + className + ". It does not implement " + Plugin.class.getName());
+            throw createClassDoesNotImplementPlugin(pluginSpecification, aClass);
         } catch (ClassNotFoundException | NoClassDefFoundError e) {
-            throw new CucumberException("Couldn't load plugin class: " + className, e);
+            throw createCouldNotLoadClass(pluginSpecification, pluginName, e);
         }
+    }
+
+    private static IllegalArgumentException createPluginIsNotCompatible(String pluginSpecification) {
+        return new IllegalArgumentException(invalidPluginMessage(pluginSpecification,
+            "This plugin is not compatible with this version of Cucumber"));
+    }
+
+    private static IllegalArgumentException createClassDoesNotImplementPlugin(
+            String pluginSpecification,
+            Class<?> pluginClass
+    ) {
+        return new IllegalArgumentException(invalidPluginMessage(pluginSpecification,
+            "'" + pluginClass.getName() + "' does not implement '" + Plugin.class.getName() + "'"));
+    }
+
+    private static IllegalArgumentException createCouldNotLoadClass(
+            String pluginSpecification, String className,
+            Throwable e
+    ) {
+        return new IllegalArgumentException(
+            invalidPluginMessage(pluginSpecification, "Could not load plugin class '" + className + "'"), e);
+    }
+
+    private static String invalidPluginMessage(String pluginSpecification, String problem) {
+        return "The plugin specification '" + pluginSpecification + "' has a problem:\n" +
+                "\n" +
+                problem + ".\n" +
+                "\n" +
+                "Plugin specifications should have the format of PLUGIN[:[PATH|[URI [OPTIONS]]]\n" +
+                "\n" +
+                "Valid values for PLUGIN are: " + PLUGIN_CLASSES.keySet().stream().sorted()
+                        .collect(joining(", "))
+                + "\n" +
+                "\n" +
+                "PLUGIN can also be a fully qualified class name, allowing registration of 3rd party plugins. " +
+                "The 3rd party plugin must implement " + Plugin.class.getName();
     }
 
     @Override